adjust menu and url links
some links were not correct. E.g. after edit the url changes with view parameter and the links in the menu don't match anymore so the correct active menu element can't be set to active.
So most of the links are now the same before and after edit.

change class a.collapse to a.has-arrow. Seems that first one isn't used anymore.